+++ src/main/kotlin/lotto/domain/Lotto.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class Lotto {
     val numbers: List<Int> = (1..45).shuffled().take(6)
