+++ src/main/kotlin/lotto/domain/LottoGame.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import lotto.util.Rank
