+++ src/test/kotlin/lotto/domain/LottoMachineTest.kt
@@ -1,6 +1,5 @@
 package lotto.domain
 
-import lotto.LottoMachine
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
